backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
redis
qdrant-client
python-dotenv
pymupdf
langchain-text-splitters
langchain-nvidia-ai-endpoints
langchain-qdrant
tiktoken